Removed ConveRTTokenizer and moved its logic into ConveRTFeaturizer. Also moved tests from test_convert_tokenizer into test_convert_featurizer and adjusted accordingly
